Using ServiceProvider that throws when getting services, rather than when creating it
